+++ app/models/spotlight/ability.rb
@@ -4,7 +4,7 @@ module Spotlight
   module Ability
     include CanCan::Ability
 
-    # rubocop:disable Metrics/AbcSize, Metrics/MethodLength, Metrics/PerceivedComplexity
+    # rubocop:disable Metrics/AbcSize, Metrics/MethodLength, Metrics/PerceivedComplexity, Metrics/CyclomaticComplexity
     def initialize(user)
       user ||= Spotlight::Engine.user_class.new
 
@@ -16,6 +16,7 @@ module Spotlight
         else
           can :manage, :all
         end
+        cannot :manage, Spotlight::Site.instance unless user.site_admin?
       end
 
       if AccessModeService.limit_access_to_site_admins?
@@ -52,6 +53,6 @@ module Spotlight
       can :read, Spotlight::Search, published: true
       can :read, Spotlight::Language, public: true
     end
-    # rubocop:enable Metrics/AbcSize, Metrics/MethodLength, Metrics/PerceivedComplexity
+    # rubocop:enable Metrics/AbcSize, Metrics/MethodLength, Metrics/PerceivedComplexity, Metrics/CyclomaticComplexity
   end
 end