+++ config/environments/production.rb
@@ -1,4 +1,4 @@
-Rails.application.configure do
+Rails.application.configure do # rubocop:disable Metrics/BlockLength
   # Settings specified here will take precedence over those in config/application.rb.
 
   # Code is not reloaded between requests.
@@ -61,12 +61,12 @@ Rails.application.configure do
   config.action_mailer.perform_deliveries = true
   config.action_mailer.raise_delivery_errors = false # Not raising errors as it causes jobs to fail AND repeat for notification errors
   config.action_mailer.smtp_settings = {
-      :address => ENV["SMTP_ADDRESS"],
-      :port => ENV["SMTP_PORT"],
-      :user_name => ENV["SMTP_USERNAME"],
-      :password => ENV["SMTP_PASSWORD"],
-      :authentication => :login,
-      :enable_starttls_auto => true
+    address: ENV["SMTP_ADDRESS"],
+    port: ENV["SMTP_PORT"],
+    user_name: ENV["SMTP_USERNAME"],
+    password: ENV["SMTP_PASSWORD"],
+    authentication: :login,
+    enable_starttls_auto: true
   }
 
   # Enable locale fallbacks for I18n (makes lookups for any locale fall back to